+++ studio/src/components/analytics/useAnalyticsQueryState.ts
@@ -68,17 +68,19 @@ export const useAnalyticsQueryState = () => {
       refreshIntervals[0]
     );
 
-    let sort = group ? {
-       id: 'totalRequests',
-       desc: true
-    }: {
-      id: 'unixTimestamp',
-      desc: true
-    }
+    let sort = group
+      ? {
+          id: "totalRequests",
+          desc: true,
+        }
+      : {
+          id: "unixTimestamp",
+          desc: true,
+        };
 
     if (query.sort) {
-      sort.id = query.sort.toString()
-      sort.desc = query.sortDir === 'desc'
+      sort.id = query.sort.toString();
+      sort.desc = query.sortDir === "desc";
     }
 
     return {
@@ -88,7 +90,7 @@ export const useAnalyticsQueryState = () => {
       dateRange,
       page,
       refreshInterval: refreshIntervalObject,
-      sort
+      sort,
     };
   }, [query]);
 };